api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
duckdb==1.1.2
pandas==2.2.2
python-dateutil==2.9.0.post0
pydantic
python-dotenv
google-generativeai
